+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0 - 2025-06-20
+
+- Allow using the `.optional()` modifier on `.nullable()` types
+- Add detailed parser descriptions to documentation
+- Add documentation for [nullable and optional modifier](doc/modifiers/nullability.md)
+- Add new type for DateTime
+
 ## 0.3.0 - 2025-06-16
 
 - Update README.md